backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
itsdangerous==2.2.0
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
av==12.0.0
yt-dlp==2025.10.14
numpy==1.26.4
sqlalchemy==2.0.27
pymysql==1.1.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
urllib3<2.0.0; python_version < "3.10"
